backend/requirements.txt
apns2==0.3.0
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.23.0
astroid==1.5.3
attrs==17.3.0
autobahn==17.10.1
Automat==0.6.0
certifi==2017.11.5
cffi==1.11.2
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.3
daphne==1.3.0
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.34.0
django-filter==1.1.0
django-push-notifications==1.5.0
django-rest-auth==0.9.2
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
iso8601==0.1.12
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack-python==0.4.8
oauthlib==2.0.6
pkg-resources==0.0.0
promise==2.1
pycparser==2.18
PyJWT==1.5.3
pylint==1.7.4
python3-openid==3.1.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
Rx==1.6.0
singledispatch==3.4.0.3
six==1.11.0
social-auth-core==1.5.0
Twisted==17.9.0
txaio==2.8.2
typing==3.6.2
urllib3==1.22
websocket-client==0.44.0
wrapt==1.10.11
zope.interface==4.4.3
